requirements.txt
# Distill CLI requirements

# Core (standard library)
# argparse, sys, json, os, re

# For YAML output format
pyyaml>=6.0

# For development
pytest>=7.0.0